rework ssh connector + fixes
