+++ src/server/JobRoutes.ts
@@ -9,7 +9,7 @@ import {
   UpdateJobBodySchema,
 } from "../definitions";
 import * as Helper from "../helpers/Helper";
-import JobUtil from "../helpers/JobUtil";
+import { validateJob } from "../helpers/JobUtil";
 import { Job } from "../models";
 import dataSource from "../utils/DB";
 
@@ -245,7 +245,7 @@ jobRouter.post("/:jobId/submit", authMiddleWare, async function (req, res) {
   
   try {
     // validate job and push it to the job queue
-    JobUtil.validateJob(job);
+    validateJob(job);
     await supervisor.pushJobToQueue(job);
   
     // update status of the job
