+++ src/server/UserRoutes.ts
@@ -8,7 +8,7 @@ import {
 } from "../../configs/config";
 import { GlobusClient } from "../helpers/GlobusTransferUtil";
 import * as Helper from "../helpers/Helper";
-import JobUtil from "../helpers/JobUtil";
+import { getUserSlurmUsage } from "../helpers/JobUtil";
 import { Job } from "../models";
 import dataSource from "../utils/DB";
 
@@ -154,7 +154,7 @@ userRouter.get("/slurm-usage", authMiddleWare, async (req, res) => {
   
   // get all jobs associated with user, then aggregate, then return that
   res.json(
-    await JobUtil.getUserSlurmUsage(res.locals.username as string, true)
+    await getUserSlurmUsage(res.locals.username as string, true)
   );
 });
